pyproject.toml
[project]
name = "Coral-PowerPointMCP-Agent"
version = "0.1.0"
description = "PowerPoint presentation agent capable of creating comprehensive presentations with tables, charts, shapes, and professional design elements based on user requests."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai[logfire]>=0.4.2",
    "pydantic-ai-slim[groq,mcp]>=0.4.2",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
